chore: Moves some code around, improves naming here and there
