+++ src/app.js
@@ -2,13 +2,13 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const swaggerUi = require('swagger-ui-express');
 const app = express();
-const config = require('../config');
-const { version } = require('../../package.json');
+const config = require('./config');
+const { version } = require('../package.json');
 
-const { validateWhiteList } = require('../helpers/validators');
+const { validateWhiteList } = require('./helpers/validators');
 const { hotelsRouter } = require('./routes/hotels/hotels');
 
-app.use('/docs', swaggerUi.serve, swaggerUi.setup(require('../../docs/swagger.json')));
+app.use('/docs', swaggerUi.serve, swaggerUi.setup(require('../docs/swagger.json')));
 app.use(bodyParser.json());
 app.use('/*', validateWhiteList);
 app.use(hotelsRouter);
@@ -19,7 +19,9 @@ app.use((err, req, res, next) => {
     short: err.short,
     long: err.long,
   });
-  if (config.get('log')) console.error(err);
+  if (config.get('log')) {
+    console.error(err);
+  }
 });
 
 app.use('/', (req, res) => {
