+++ src/configuration.json
@@ -4,7 +4,7 @@
     "gasMargin": 1.5,
     "indexAddress": "0x09C0AECBA9BBEBCCDAAF3FE8473591A69C4C11BE",
     "port": 3000,
-    "privateKeyDir": "keys/mainKey.json",
+    "privateKeyFile": "keys/mainKey.json",
     "web3Provider": "https://ropsten.infura.io/WKNyJ0kClh8Ao5LdmO7z",
     "log": true,
     "whiteList": []
@@ -13,7 +13,7 @@
     "gasMargin": 2,
     "indexAddress": "0x8c2373842d5ea4ce4baf53f4175e5e42a364c59c",
     "port": 3000,
-    "privateKeyDir": "keys/mainKey.json",
+    "privateKeyFile": "keys/mainKey.json",
     "web3Provider": "http://localhost:8545",
     "log": true,
     "whiteList": []
