+++ src/index.js
@@ -1,5 +1,5 @@
-const { app } = require('./service');
-const config = require('../config');
+const { app } = require('./app');
+const config = require('./config');
 
 app.listen(config.get('port'), () => {
   console.log(`WT API AT ${config.get('port')}!`);
