+++ test/utils/hooks.js
@@ -2,10 +2,10 @@
 /* eslint-disable no-unused-expressions */
 const { expect } = require('chai');
 const fetch = require('node-fetch');
-const config = require('../src/config');
-const { PASSWORD_HEADER } = require('../src/helpers/validators');
+const config = require('../../src/config');
+const { PASSWORD_HEADER } = require('../../src/helpers/validators');
+const { app } = require('../../src/app');
 
-const { app } = require('../src/srv/service');
 const lifData = require('@windingtree/lif-token/build/contracts/LifTokenTest.json');
 
 const gasMargin = 1.5;
@@ -20,7 +20,7 @@ const Before = () => (
   before(async function () {
     config.set('log', false);
     config.set('password', 'test123');
-    config.set('privateKeyDir', 'keys/test.json');
+    config.set('privateKeyFile', 'keys/test.json');
     const wallet = await config.get('web3provider').web3.eth.accounts.wallet.create(3);
     const createdAccounts = await config.get('web3provider').web3.eth.getAccounts();
     fundingSource = createdAccounts[0];
@@ -32,9 +32,11 @@ const Before = () => (
     await config.get('web3provider').accounts.fundAccount(fundingSource, config.get('user'), '50');
   })
 );
+
 const BeforeEach = () => (
   beforeEach(async function () {
     config.set('whiteList', ['127.0.0.1']);
+    // TODO deploy manually from here,  deployIndex will not exist in next wt-js-libs iteration
     index = await config.get('web3provider').deploy.deployIndex(daoAccount, gasMargin);
     expect(index._address).to.not.equal(addressZero);
     config.set('indexAddress', index._address);
@@ -44,6 +46,7 @@ const BeforeEach = () => (
     await deployLifContract(daoAccount, config.get('user'), index);
   })
 );
+
 const AfterEach = () => (
   afterEach(async function () {
     return server.close();
